Working on Lobby List for map markers
Tried to test Siyao's code but it wasn't working, so I modified Blong's
code for parsing a JSON Array. The code is nearly working but I think
there is a problem in that the background task is not finishing before
the map is being created so it is now displaying the markers.